center-y

+++ src/project_iso.c
@@ -6,7 +6,7 @@
 /*   By: abazzoun <abazzoun@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/09/05 10:47:12 by abazzoun          #+#    #+#             */
-/*   Updated: 2025/09/17 23:47:03 by abazzoun         ###   ########.fr       */
+/*   Updated: 2025/09/21 02:10:20 by abazzoun         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -45,8 +45,8 @@ t_point2d	project_iso(int x, int y, t_cell cell, t_props *props)
 	p.z = cell.z * props->scale * props->scale_z;
 	p = rotate_point(p.x, p.y, p.z, props);
 	angle = M_PI / 6;
-	proj.x = (int)((p.x - p.y) * cos(angle)) + props->offset_x;
-	proj.y = (int)((p.x + p.y) * sin(angle) - p.z) + props->offset_y;
+	proj.x = ((p.x - p.y) * cos(angle)) + props->offset_x;
+	proj.y = ((p.x + p.y) * sin(angle) - p.z) + props->offset_y;
 	proj.color = cell.color;
 	return (proj);
 }
